agent_with_tool/requirements.txt
openai==0.27.8
langchain==0.0.266
streamlit==1.26.0
google-api-python-client==2.97.0
youtube_transcript_api==0.6.1
spotipy==2.23.0
duckduckgo-search==3.8.5
tweepy==4.13.0
google-cloud-bigquery==3.11.4
audio_recorder_streamlit